Release v6.0.2-debug.2: Optimise Screenshot Logic
Removed screen_metrics.dart as the device pixel ratio (DPR) does not change when the orientation of an Android device changes. The DPR is a hardware characteristic of the display and is fixed for a given device. It represents the ratio between physical pixels and logical (or CSS) pixels.

Added a static variable to get the DPR.

Added debug logs

Optimised render_tree.dart and ue_snapshotter.dart